Move preparers, transformers, and pycon utils into shell_evaluator package
Converts shell_evaluator from a single module into a package, nesting
source_preparer, result_transformer, and _pycon as sub-modules. Also
adds PyconSourcePreparer and PyconResultTransformer implementations, and
creates test_pycon_shell_evaluator.py for pycon-specific tests.
